% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMfunction.R
\name{RFMfunction}
\alias{RFMfunction}
\title{Caluclate RFM}
\usage{
RFMfunction(data, weight_recency = 1, weight_frequency = 1,
  weight_monetary = 1)
}
\arguments{
\item{data}{- a data table}

\item{r}{- weight recency}

\item{f}{- weight frequency}

\item{m}{- weight monetary}
}
\value{
returns a data table with the weighted rfm scores for each customer
}
\description{
Caluclate RFM
}
\details{
\code{data} contains ...
}
